+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## next
+
+- Add `isNumberOrString` check
+
 ## 1.0.0
 
 - initial Release